finished login tests

+++ lib/dsa/accounts.ex
@@ -2,64 +2,45 @@ defmodule Dsa.Accounts do
   @moduledoc """
   The Accounts context.
   """
-  import Ecto.Query, warn: false
 
+  import Ecto.Query, warn: false
   alias Dsa.Repo
-  alias Dsa.Accounts.{UserCredential, Group, User, UserToken, UserNotifier}
-  alias Dsa.Characters.Character
+  alias Dsa.Accounts.{User, UserToken, UserNotifier}
 
   ## Database getters
 
   @doc """
-  Gets a credential by user_id.
+  Gets a user by email.
 
   ## Examples
 
-      iex> get_credential(1)
+      iex> get_user_by_email("foo@example.com")
       %User{}
 
-      iex> get_credential(-1)
+      iex> get_user_by_email("unknown@example.com")
       nil
 
   """
-  def get_credential(id) when is_integer(id) do
-    Repo.get(UserCredential, id)
+  def get_user_by_email(email) when is_binary(email) do
+    Repo.get_by(User, email: email)
   end
 
   @doc """
-  Gets a credential by email.
+  Gets a user by email and password.
 
   ## Examples
 
-      iex> get_credential_by_email("foo@example.com")
+      iex> get_user_by_email_and_password("foo@example.com", "correct_password")
       %User{}
 
-      iex> get_credential_by_email("unknown@example.com")
+      iex> get_user_by_email_and_password("foo@example.com", "invalid_password")
       nil
 
   """
-  def get_credential_by_email(email) when is_binary(email) do
-    Repo.get_by(UserCredential, email: email)
-  end
-
-  @doc """
-  Gets a credential by email and password.
-
-  ## Examples
-
-      iex> get_credential_by_email_and_password("foo@example.com", "correct_password")
-      %User{}
-
-      iex> get_credential_by_email_and_password("foo@example.com", "invalid_password")
-      nil
-
-  """
-  def get_credential_by_email_and_password(email, password)
+  def get_user_by_email_and_password(email, password)
       when is_binary(email) and is_binary(password) do
-    credential = Repo.get_by(UserCredential, email: email)
-    if UserCredential.valid_password?(credential, password) do
-      credential
-    end
+    user = Repo.get_by(User, email: email)
+    if User.valid_password?(user, password), do: user
   end
 
   @doc """
@@ -81,22 +62,20 @@ defmodule Dsa.Accounts do
   ## User registration
 
   @doc """
-  Registers a user by creating a user and an associated credential.
-  Returns the credential.
+  Registers a user.
 
   ## Examples
 
       iex> register_user(%{field: value})
-      {:ok, %Credential{}}
+      {:ok, %User{}}
 
       iex> register_user(%{field: bad_value})
       {:error, %Ecto.Changeset{}}
 
   """
   def register_user(attrs) do
-    %UserCredential{}
-    |> UserCredential.registration_changeset(attrs)
-    |> Ecto.Changeset.cast_assoc(:user, with: &User.changeset/2)
+    %User{}
+    |> User.registration_changeset(attrs)
     |> Repo.insert()
   end
 
@@ -105,14 +84,12 @@ defmodule Dsa.Accounts do
 
   ## Examples
 
-      iex> change_user_registration(credential)
-      %Ecto.Changeset{data: %Credential{}}
+      iex> change_user_registration(user)
+      %Ecto.Changeset{data: %User{}}
 
   """
-  def change_user_registration(%UserCredential{} = credential, attrs \\ %{}) do
-    credential
-    |> UserCredential.registration_changeset(attrs, hash_password: false)
-    |> Ecto.Changeset.cast_assoc(:user, with: &User.changeset/2)
+  def change_user_registration(%User{} = user, attrs \\ %{}) do
+    User.registration_changeset(user, attrs, hash_password: false)
   end
 
   ## Settings
@@ -123,14 +100,11 @@ defmodule Dsa.Accounts do
   ## Examples
 
       iex> change_user_email(user)
-      %Ecto.Changeset{data: %Credential{}}
+      %Ecto.Changeset{data: %User{}}
 
   """
   def change_user_email(user, attrs \\ %{}) do
-    user
-    |> Repo.preload(:credential)
-    |> Map.get(:credential)
-    |> UserCredential.email_changeset(attrs)
+    User.email_changeset(user, attrs)
   end
 
   @doc """
@@ -148,10 +122,8 @@ defmodule Dsa.Accounts do
   """
   def apply_user_email(user, password, attrs) do
     user
-    |> Repo.preload(:credential)
-    |> Map.get(:credential)
-    |> UserCredential.email_changeset(attrs)
-    |> UserCredential.validate_current_password(password)
+    |> User.email_changeset(attrs)
+    |> User.validate_current_password(password)
     |> Ecto.Changeset.apply_action(:update)
   end
 
@@ -161,27 +133,24 @@ defmodule Dsa.Accounts do
   If the token matches, the user email is updated and the token is deleted.
   The confirmed_at date is also updated to the current time.
   """
-  def update_credential_email(credential, token) do
-    context = "change:#{credential.email}"
+  def update_user_email(user, token) do
+    context = "change:#{user.email}"
 
     with {:ok, query} <- UserToken.verify_change_email_token_query(token, context),
          %UserToken{sent_to: email} <- Repo.one(query),
-         {:ok, _} <- Repo.transaction(credential_email_multi(credential, email, context)) do
+         {:ok, _} <- Repo.transaction(user_email_multi(user, email, context)) do
       :ok
     else
       _ -> :error
     end
   end
 
-  defp credential_email_multi(credential, email, context) do
-    changeset =
-      credential
-      |> UserCredential.email_changeset(%{email: email})
-      |> UserCredential.confirm_changeset()
+  defp user_email_multi(user, email, context) do
+    changeset = user |> User.email_changeset(%{email: email}) |> User.confirm_changeset()
 
     Ecto.Multi.new()
-    |> Ecto.Multi.update(:credential, changeset)
-    |> Ecto.Multi.delete_all(:tokens, UserToken.user_and_contexts_query(credential.user_id, [context]))
+    |> Ecto.Multi.update(:user, changeset)
+    |> Ecto.Multi.delete_all(:tokens, UserToken.user_and_contexts_query(user, [context]))
   end
 
   @doc """
@@ -206,12 +175,12 @@ defmodule Dsa.Accounts do
 
   ## Examples
 
-      iex> change_user_password(credential)
-      %Ecto.Changeset{data: %Credential{}}
+      iex> change_user_password(user)
+      %Ecto.Changeset{data: %User{}}
 
   """
-  def change_user_password(credential, attrs \\ %{}) do
-    UserCredential.password_changeset(credential, attrs, hash_password: false)
+  def change_user_password(user, attrs \\ %{}) do
+    User.password_changeset(user, attrs, hash_password: false)
   end
 
   @doc """
@@ -219,22 +188,22 @@ defmodule Dsa.Accounts do
 
   ## Examples
 
-      iex> update_user_password(credential, "valid password", %{password: ...})
+      iex> update_user_password(user, "valid password", %{password: ...})
       {:ok, %User{}}
 
-      iex> update_user_password(credential, "invalid password", %{password: ...})
+      iex> update_user_password(user, "invalid password", %{password: ...})
       {:error, %Ecto.Changeset{}}
 
   """
-  def update_user_password(credential, password, attrs) do
+  def update_user_password(user, password, attrs) do
     changeset =
-      credential
-      |> UserCredential.password_changeset(attrs)
-      |> UserCredential.validate_current_password(password)
+      user
+      |> User.password_changeset(attrs)
+      |> User.validate_current_password(password)
 
     Ecto.Multi.new()
-    |> Ecto.Multi.update(:credential, changeset)
-    |> Ecto.Multi.delete_all(:tokens, UserToken.user_and_contexts_query(credential.user_id, :all))
+    |> Ecto.Multi.update(:user, changeset)
+    |> Ecto.Multi.delete_all(:tokens, UserToken.user_and_contexts_query(user, :all))
     |> Repo.transaction()
     |> case do
       {:ok, %{user: user}} -> {:ok, user}
@@ -247,23 +216,18 @@ defmodule Dsa.Accounts do
   @doc """
   Generates a session token.
   """
-  def generate_user_session_token(credential) do
-    {token, user_token} = UserToken.build_session_token(credential)
+  def generate_user_session_token(user) do
+    {token, user_token} = UserToken.build_session_token(user)
     Repo.insert!(user_token)
     token
   end
 
   @doc """
-  Gets the user with the given signed token and preloads essential information.
+  Gets the user with the given signed token.
   """
   def get_user_by_session_token(token) do
     {:ok, query} = UserToken.verify_session_token_query(token)
-
-    character_query = from(c in Character, select: map(c, [:id, :name]))
-
-    query
-    |> Repo.one()
-    |> Repo.preload(characters: character_query)
+    Repo.one(query)
   end
 
   @doc """
@@ -288,17 +252,14 @@ defmodule Dsa.Accounts do
       {:error, :already_confirmed}
 
   """
-  def deliver_user_confirmation_instructions(%UserCredential{} = credential, confirmation_url_fun)
+  def deliver_user_confirmation_instructions(%User{} = user, confirmation_url_fun)
       when is_function(confirmation_url_fun, 1) do
-
-    if credential.confirmed_at do
+    if user.confirmed_at do
       {:error, :already_confirmed}
     else
-      {encoded_token, user_token} = UserToken.build_email_token(credential, "confirm")
+      {encoded_token, user_token} = UserToken.build_email_token(user, "confirm")
       Repo.insert!(user_token)
-      UserNotifier.deliver_confirmation_instructions(
-        credential.email,
-        confirmation_url_fun.(encoded_token))
+      UserNotifier.deliver_confirmation_instructions(user, confirmation_url_fun.(encoded_token))
     end
   end
 
@@ -311,18 +272,17 @@ defmodule Dsa.Accounts do
   def confirm_user(token) do
     with {:ok, query} <- UserToken.verify_email_token_query(token, "confirm"),
          %User{} = user <- Repo.one(query),
-         {:ok, %{user: user}} <- Repo.transaction(confirm_user_multi(user.credential)) do
+         {:ok, %{user: user}} <- Repo.transaction(confirm_user_multi(user)) do
       {:ok, user}
     else
       _ -> :error
     end
   end
 
-  defp confirm_user_multi(credential) do
+  defp confirm_user_multi(user) do
     Ecto.Multi.new()
-    |> Ecto.Multi.update(:credential, UserCredential.confirm_changeset(credential))
-    |> Ecto.Multi.delete_all(:tokens,
-        UserToken.user_and_contexts_query(credential.user_id, ["confirm"]))
+    |> Ecto.Multi.update(:user, User.confirm_changeset(user))
+    |> Ecto.Multi.delete_all(:tokens, UserToken.user_and_contexts_query(user, ["confirm"]))
   end
 
   ## Reset password
@@ -338,7 +298,6 @@ defmodule Dsa.Accounts do
   """
   def deliver_user_reset_password_instructions(%User{} = user, reset_password_url_fun)
       when is_function(reset_password_url_fun, 1) do
-    user = Repo.preload(user, :credential)
     {encoded_token, user_token} = UserToken.build_email_token(user, "reset_password")
     Repo.insert!(user_token)
     UserNotifier.deliver_reset_password_instructions(user, reset_password_url_fun.(encoded_token))
@@ -366,62 +325,25 @@ defmodule Dsa.Accounts do
   end
 
   @doc """
-  Resets the credential's password.
+  Resets the user password.
 
   ## Examples
 
-      iex> reset_credential_password(credential, %{password: "new long password", password_confirmation: "new long password"})
+      iex> reset_user_password(user, %{password: "new long password", password_confirmation: "new long password"})
       {:ok, %User{}}
 
-      iex> reset_credential_password(credential, %{password: "valid", password_confirmation: "not the same"})
+      iex> reset_user_password(user, %{password: "valid", password_confirmation: "not the same"})
       {:error, %Ecto.Changeset{}}
 
   """
-  def reset_credential_password(credential, attrs) do
+  def reset_user_password(user, attrs) do
     Ecto.Multi.new()
-    |> Ecto.Multi.update(:user, UserCredential.password_changeset(credential, attrs))
-    |> Ecto.Multi.delete_all(:tokens, UserToken.user_and_contexts_query(credential.user_id, :all))
+    |> Ecto.Multi.update(:user, User.password_changeset(user, attrs))
+    |> Ecto.Multi.delete_all(:tokens, UserToken.user_and_contexts_query(user, :all))
     |> Repo.transaction()
     |> case do
       {:ok, %{user: user}} -> {:ok, user}
       {:error, :user, changeset, _} -> {:error, changeset}
     end
-  end
-
-  def list_groups do
-    from(g in Group, preload: [
-      master: ^from(u in User, select: u.username),
-      users: ^from(u in User, select: u.id)
-    ])
-    |> Repo.all()
-  end
-
-  def get_group!(id), do: Repo.get!(Group, id)
-
-  def change_group(%Group{} = group, attrs \\ %{}), do: Group.changeset(group, attrs)
-
-  def create_group(%User{} = user, attrs) do
-    %Group{}
-    |> Group.changeset(attrs)
-    |> Ecto.Changeset.put_assoc(:master, user)
-    |> Repo.insert()
-  end
-
-  def delete_group(%Group{} = group), do: Repo.delete(group)
-
-  def join_group(%User{} = user, %Group{} = group) do
-    user
-    |> Repo.preload(:group)
-    |> Ecto.Changeset.change()
-    |> Ecto.Changeset.put_assoc(:group, group)
-    |> Repo.update()
-  end
-
-  def leave_group(%User{} = user) do
-    user
-    |> Repo.preload(:group)
-    |> Ecto.Changeset.change()
-    |> Ecto.Changeset.put_assoc(:group, nil)
-    |> Repo.update()
   end
 end
